feat(doctrine) Added Document table and entity and updated other tables
