+++ translations/validators/validators.en.yaml
@@ -1,10 +1,11 @@
-article.title.not_blank: "The title of an article can not be blank"
-article.title.length: "The title of an article can not exceed {{ limit }} characters"
-article.description.not_blank: "The description of an article can not be blank"
-article.description.length: "The description of an article can not exceed {{ limit }} characters"
-article.content.not_blank: "The content of an article can not be blank"
-article.views.positive_of_zero: "The number of views must be positive or zero"
-article.slug.length: "The slug of an article can not exceed {{ limit }} characters"
+document.title.not_blank: "The title of an document can not be blank"
+document.title.length: "The title of an document can not exceed {{ limit }} characters"
+document.description.not_blank: "The description of an document can not be blank"
+document.description.length: "The description of an document can not exceed {{ limit }} characters"
+document.content.not_blank: "The content of an document can not be blank"
+document.views.positive_of_zero: "The number of views must be positive or zero"
+document.slug.length: "The slug of an document can not exceed {{ limit }} characters"
+
 article.thumbnail.not_null: "The thumbnail of an article must be filled in"
 article.thumbnail.max_size: "The thumbnail of an article can not exceed {{ limit }}{{ suffix }}"
 article.thumbnail.mime_type: "Wrong thumbnail format (jpeg, jpg, png)"
