+++ translations/validators/validators.fr.yaml
@@ -1,10 +1,11 @@
-article.title.not_blank: "Le titre d'un article ne pas être vide"
-article.title.length: "Le titre d'un article ne peut pas dépasser {{ limit }} caractères"
-article.description.not_blank: "La description d'un article ne pas être vide"
-article.description.length: "La description d'un article ne peut pas dépasser {{ limit }} caractères"
-article.content.not_blank: "La contenu d'un article ne pas être vide"
-article.views.positive_of_zero: "Le nombre de vues d'un article doit être supérieur ou égale à 0"
-article.slug.length: "Le slug d'un article ne peut pas dépasser {{ limit }} caractères"
+document.title.not_blank: "Le titre d'un document ne pas être vide"
+document.title.length: "Le titre d'un document ne peut pas dépasser {{ limit }} caractères"
+document.description.not_blank: "La description d'un document ne pas être vide"
+document.description.length: "La description d'un document ne peut pas dépasser {{ limit }} caractères"
+document.content.not_blank: "La contenu d'un document ne pas être vide"
+document.views.positive_of_zero: "Le nombre de vues d'un document doit être supérieur ou égale à 0"
+document.slug.length: "Le slug d'un document ne peut pas dépasser {{ limit }} caractères"
+
 article.thumbnail.not_null: "L'image d'un article doit être renseignée"
 article.thumbnail.max_size: "L'image ne peut pas dépasser {{ limit }}{{ suffix }}"
 article.thumbnail.mime_type: "Mauvais format d'image (jpeg, jpg, png)"
